backend/requirements.txt
# 核心后端框架
fastapi==0.115.0
uvicorn[standard]==0.30.6

# 数据库与ORM
sqlalchemy==2.0.34
psycopg2-binary==2.9.9
asyncpg==0.30.0

# 配置与数据验证
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2

# NLP与向量化（兼容Python 3.9）
sentence-transformers==3.1.1
numpy==1.26.4
scikit-learn==1.5.2

# 数据处理
pandas==2.2.3
openpyxl==3.1.5

# 缓存（修正版本号！7.4.0不存在）
redis==5.0.8

# 异步工具
httpx==0.27.2

# 爬虫：动态页兜底（可选，未安装时仅用静态抓取）
playwright==1.49.0
